finish

+++ app/Http/Controllers/TasksController.php
@@ -15,12 +15,21 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();//
-        return view('tasks.index',[
-            'tasks'=>$tasks,
-        ]);
-    }
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('users.show', $data);
+        }else {
+            return view('welcome');
+        }
+    }
     /**
      * Show the form for creating a new resource.
      *
@@ -43,17 +52,32 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request,[
-            'status'=>'required|max:10',
-            'content'=>'required|max:10',
-            ]);
-            
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        $id=\Auth::id();
+        
+        $this->validate($request, [
+            'content' => 'required|max:191',
+            'status'=>'required|max:191'
+        ]);
 
-        return redirect('/');//
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status'=>$request->status,
+        ]);
+        
+
+        return redirect()->route("users.show", ["id"=>$id]);
+        
+        // $this->validate($request,[
+        //     'status'=>'required|max:10',
+        //     'content'=>'required|max:10',
+        //     ]);
+            
+        // $task = new Task;
+        // $task->status = $request->status;
+        // $task->content = $request->content;
+        // $task->save();
+
+        // return redirect('/');//
     }
 
     /**
@@ -64,11 +88,19 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::find($id);
+         $task = Task::find($id);
+        
+        if (\Auth::id() === $task->user_id) {
+        
+       
         
         return view('tasks.show', [
             'task' => $task,
             ]);
+        }
+        else {
+            return redirect("/");
+        }
     }
 
     /**
@@ -78,12 +110,24 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
-    {
-        $task = Task::find($id);
+    
+        
+        // if (\Auth::id() === $task->user_id) {
 
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);//
+        // return view('tasks.edit', [
+        //     'task' => $task,
+        // ]);
+        {
+        $task = Task::find($id);
+        
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit',['task'=>$task]);
+            
+        }
+        
+        else {
+            return redirect("/");
+        }
     }
 
     /**
@@ -117,8 +161,19 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
 
-        return redirect('/');//
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+
+            
+        }
+        
+            return redirect("/");
+        
+        
+        // $task = Task::find($id);
+        // $task->delete();
+
+        // return redirect('/');//
     }
 }
